// packages/schemas/src/nfe/evento-cancelamento/PL_010d.ts
export type TEvento_infEvento = {
  /**
   * Identificador da TAG a ser assinada, a regra de formação do Id é:
   * “ID” + tpEvento +  chave da NF-e + nSeqEvento
   * @attribute pattern `ID[0-9]{12}[0-9A-Z]{12}[0-9]{28}`
   */
  Id: string;
  /**
   * Código do órgão de recepção do Evento. Utilizar a Tabela do IBGE extendida, utilizar 90 para identificar o Ambiente Nacional
   * xsd:TCOrgaoIBGE
   */
  cOrgao: TCOrgaoIBGE;
  /**
   * Identificação do Ambiente:
   * 1 - Produção
   * 2 - Homologação
   * xsd:TAmb
   */
  tpAmb: TAmb;
  /**
   * Chave de Acesso da NF-e vinculada ao evento
   * xsd:TChNFe, pattern `[0-9]{6}[0-9A-Z]{12}[0-9]{26}`
   */
  chNFe: string;
  /**
   * Data e Hora do Evento, formato UTC (AAAA-MM-DDThh:mm:ssTZD, onde TZD = +hh:mm ou -hh:mm)
   * xsd:TDateTimeUTC
   */
  dhEvento: string;
  /**
   * Tipo do Evento
   * pattern `[0-9]{6}`
   */
  tpEvento: string;
  /**
   * Seqüencial do evento para o mesmo tipo de evento.  Para maioria dos eventos será 1, nos casos em que possa existir mais de um evento, como é o caso da carta de correção, o autor do evento deve numerar de forma seqüencial.
   * pattern `[1-9][0-9]{0,1}`
   */
  nSeqEvento: string;
  /** Versão do Tipo do Evento */
  verEvento: string;
  /** Schema XML de validação do evento do cancelamento 1101111 */
  detEvento: TEvento_infEvento_detEvento;
  /** Grupo de Informação do Provedor de Assinatura e Autorização */
  infPAA?: TEvento_infEvento_infPAA;
} & (
  ({
  /**
   * CNPJ
   * xsd:TCnpjOpc, tamanho 0..14, pattern `[0-9]{0}|[0-9A-Z]{12}[0-9]{2}`
   */
  CNPJ: string;
  CPF?: never;
})
  | ({
  /**
   * CPF
   * xsd:TCpf, pattern `[0-9]{11}`
   */
  CPF: string;
  CNPJ?: never;
})
);
// ...
export type SignatureType = {
  /** @attribute */
  Id?: string;
  SignedInfo: SignedInfoType;
  SignatureValue: SignatureValueType;
  KeyInfo: KeyInfoType;
};

/**
 * Tipo Evento
 * xsd: TEvento
 */
export type TEvento = {
  /** @attribute xsd:TVerEvento, pattern `1\.00` */
  versao: string;
  infEvento: TEvento_infEvento;
  Signature: SignatureType;
};

/**
 * Tipo Lote de Envio
 * xsd: TEnvEvento
 */
export type TEnvEvento = {
  /** @attribute xsd:TVerEnvEvento, pattern `1\.00` */
  versao: string;
  /** pattern `[0-9]{1,15}` */
  idLote: string;
  /** ocorre 1..20 */
  evento: TEvento[];
};
// ...
export const TEvento_infEvento: ComplexType<TEvento_infEvento> = { id: "TEvento.infEvento", ns: "http://www.portalfiscal.inf.br/nfe", a: [{ a: "Id", t: st$1, r: 1 }], c: { g: "s", i: [{ e: "cOrgao", t: st_TCOrgaoIBGE }, { e: "tpAmb", t: st_TAmb }, { g: "c", i: [{ e: "CNPJ", t: st_TCnpjOpc }, { e: "CPF", t: st_TCpf }] }, { e: "chNFe", t: st_TChNFe }, { e: "dhEvento", t: st_TDateTimeUTC }, { e: "tpEvento", t: st$2 }, { e: "nSeqEvento", t: st$3 }, { e: "verEvento", t: st$4 }, { e: "detEvento", t: TEvento_infEvento_detEvento }, { e: "infPAA", t: TEvento_infEvento_infPAA, n: 0 }] } };
// ...
export const SignatureType: ComplexType<SignatureType> = { id: "SignatureType", ns: "http://www.w3.org/2000/09/xmldsig#", a: [{ a: "Id", t: st$8 }], c: { g: "s", i: [{ e: "SignedInfo", t: SignedInfoType }, { e: "SignatureValue", t: SignatureValueType }, { e: "KeyInfo", t: KeyInfoType }] } };
export const TEvento: ComplexType<TEvento> = { id: "TEvento", ns: "http://www.portalfiscal.inf.br/nfe", a: [{ a: "versao", t: st_TVerEvento, r: 1 }], c: { g: "s", i: [{ e: "infEvento", t: TEvento_infEvento }, { e: "Signature", t: SignatureType, ns: "http://www.w3.org/2000/09/xmldsig#" }] } };
export const TEnvEvento: ComplexType<TEnvEvento> = { id: "TEnvEvento", ns: "http://www.portalfiscal.inf.br/nfe", a: [{ a: "versao", t: st_TVerEnvEvento, r: 1 }], c: { g: "s", i: [{ e: "idLote", t: st$0 }, { e: "evento", t: TEvento, x: 20 }] } };
// ...
export const envEventoElement: RootElement<TEnvEvento> = { name: "envEvento", ns: "http://www.portalfiscal.inf.br/nfe", type: TEnvEvento };